+++ app/src/main/res/layout/item_produto.xml
@@ -22,6 +22,17 @@
             android:layout_height="wrap_content"
             android:layout_below="@+id/nomeProduto" />
 
+        <TextView
+            android:id="@+id/valorVenda"
+            android:layout_width="56dp"
+            android:layout_height="30dp"
+            android:layout_alignParentEnd="true"
+            android:layout_alignTop="@+id/nomeProduto"
+            android:layout_marginEnd="14dp"
+            android:text="R$"
+            android:layout_marginRight="14dp"
+            android:layout_alignParentRight="true" />
+
     </RelativeLayout>
 
 </android.support.v7.widget.CardView>
